+++ src/bvals/bvals_refine.cpp
@@ -106,7 +106,7 @@ void BoundaryValues::FillRestrictionMetadata(cell_centered_bvars::BufferCacheHos
               info(idx).allocated = v->IsAllocated();
               if (v->IsAllocated()) {
                 info(idx).fine = v->data.Get();
-                info(idx).coarse = v->vbvar->coarse_buf.Get();
+                info(idx).coarse = v->coarse_s.Get();
               }
               info(idx).refinement_op = RefinementOp_t::Restriction;
               info(idx).Nt = v->GetDim(6);